function/pyproject.toml
[project]
name = "lm-sentinel-function"
version = "0.1.0"
description = "Azure Function that routes LogicMonitor LogAlerts into Microsoft Sentinel via Logs Ingestion API"
requires-python = ">=3.12"
dependencies = [
    "azure-functions>=1.21.0",
    "azure-identity>=1.19.0",
    "azure-monitor-ingestion>=1.0.4",
    "pydantic>=2.9.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.6",
    "mypy>=1.11",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "RUF"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
